+++ projects/hipdnn/test_sdk/tests/utilities/cpu_graph_executor/LayernormGraphUtils.hpp
@@ -82,10 +82,10 @@ inline std::shared_ptr<hipdnn_frontend::graph::Graph>
     auto epsilonTensor = std::make_shared<hipdnn_frontend::graph::TensorAttributes>();
     epsilonTensor->set_uid(uid++)
         .set_name("EpsilonTensor")
-        .set_data_type(hipdnn_frontend::DataType::DOUBLE)
+        .set_data_type(hipdnn_frontend::DataType::FLOAT)
         .set_dim({1})
         .set_stride({1})
-        .set_value(hipdnn_data_sdk::utilities::LAYERNORM_DEFAULT_EPSILON);
+        .set_value(static_cast<float>(hipdnn_data_sdk::utilities::LAYERNORM_DEFAULT_EPSILON));
 
     hipdnn_frontend::graph::LayernormAttributes lnAttrs;
     lnAttrs.set_name("layernorm_fprop");
@@ -233,10 +233,10 @@ inline std::shared_ptr<hipdnn_frontend::graph::Graph>
     auto epsilonTensor = std::make_shared<hipdnn_frontend::graph::TensorAttributes>();
     epsilonTensor->set_uid(uid++)
         .set_name("EpsilonTensor")
-        .set_data_type(hipdnn_frontend::DataType::DOUBLE)
+        .set_data_type(hipdnn_frontend::DataType::FLOAT)
         .set_dim({1})
         .set_stride({1})
-        .set_value(hipdnn_data_sdk::utilities::LAYERNORM_DEFAULT_EPSILON);
+        .set_value(static_cast<float>(hipdnn_data_sdk::utilities::LAYERNORM_DEFAULT_EPSILON));
 
     hipdnn_frontend::graph::LayernormBackwardAttributes lnAttrs;
     lnAttrs.set_name("layernorm_bprop");
