+++ projects/hipdnn/test_sdk/tests/utilities/cpu_graph_executor/RMSNormGraphUtils.hpp
@@ -56,12 +56,11 @@ inline std::shared_ptr<hipdnn_frontend::graph::Graph>
     {
         // Pure runtime pass-by-value: FLOAT scalar, no baked value; resolved
         // from the variant pack at execute.
-        epsilonTensor->set_data_type(hipdnn_frontend::DataType::FLOAT)
-            .set_as_runtime_parameter();
+        epsilonTensor->set_data_type(hipdnn_frontend::DataType::FLOAT).set_as_runtime_parameter();
     }
     else
     {
-        epsilonTensor->set_data_type(hipdnn_frontend::DataType::DOUBLE).set_value(1e-5);
+        epsilonTensor->set_data_type(hipdnn_frontend::DataType::FLOAT).set_value(1e-5f);
     }
 
     hipdnn_frontend::graph::RMSNormAttributes rmsnormAttrs;
@@ -153,10 +152,10 @@ inline std::shared_ptr<hipdnn_frontend::graph::Graph>
     auto epsilonTensor = std::make_shared<hipdnn_frontend::graph::TensorAttributes>();
     epsilonTensor->set_uid(uid++)
         .set_name("EpsilonTensor")
-        .set_data_type(hipdnn_frontend::DataType::DOUBLE)
+        .set_data_type(hipdnn_frontend::DataType::FLOAT)
         .set_dim({1})
         .set_stride({1})
-        .set_value(1e-5);
+        .set_value(1e-5f);
 
     hipdnn_frontend::graph::RMSNormAttributes rmsnormAttrs;
     rmsnormAttrs.set_name("rmsnorm_fwd_bias");
